[guest,guest_bin] created GuestHandle and moved entrypoint + guest_logger + exceptions + guest function logic + guest_err
- moved entrypoint to guest_bin library. W/ this, I also moved a lot of things that were only used via the entrypoint:
-- IDT/GDT setup for exceptions.
-- guest_logger.

- w/ the entrypoint fxn from the guest lib having moved, I re-named that module to 'exit' as it still contains stuff for
aborting guest execution. I also moved OUT functions there like 'outb' and 'out32'.

- a lot of our public functions in the guest lib, like, for example, push_shared_output_data relied on a pub(crate) global
(i.e., P_PEB) being set. W/ that only ever being set in the entrypoint and w/ it not being appropriate as pub, I created a
GuestHandle struct. This struct houses a *mut HyperlightPEB that operations like push_shared_output_data can leverage.
bin-like crates can then create their own GuestHandle globals and use its functionality like self.push_shared_output_data(...).

- moved all out guest function defintion, registration and so on to the guest_bin crate.
