+++ src/hyperlight_guest/Cargo.toml
@@ -20,10 +20,7 @@ printf = [] # compile printf
 [dependencies]
 anyhow = { version = "1.0.98", default-features = false }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-buddy_system_allocator = "0.11.0"
 hyperlight-common = { workspace = true }
-spin = "0.10.0"
-log = { version = "0.4", default-features = false }
 
 [build-dependencies]
 cc = "1.2"
