+++ src/hyperlight_guest_bin/Cargo.toml
@@ -9,8 +9,11 @@ repository.workspace = true
 readme.workspace = true
 
 [dependencies]
-hyperlight-guest = { workspace = true, default-features = false}
-hyperlight-common = { workspace = true, default-features = false}
+hyperlight-guest = { workspace = true, default-features = false }
+hyperlight-common = { workspace = true, default-features = false }
+buddy_system_allocator = "0.11.0"
+log = { version = "0.4", default-features = false }
+spin = "0.10.0"
 
 [lints]
 workspace = true
