+++ src/hyperlight_guest_bin/src/guest_err.rs
@@ -14,32 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use alloc::string::ToString;
-use alloc::vec::Vec;
 use core::ffi::{c_char, CStr};
 
-use hyperlight_common::flatbuffer_wrappers::guest_error::{ErrorCode, GuestError};
+use hyperlight_common::flatbuffer_wrappers::guest_error::ErrorCode;
+use hyperlight_guest::exit::halt;
 
-use crate::entrypoint::halt;
-use crate::shared_output_data::push_shared_output_data;
+use crate::GUEST_HANDLE;
 
-pub(crate) fn write_error(error_code: ErrorCode, message: Option<&str>) {
-    let guest_error = GuestError::new(
-        error_code.clone(),
-        message.map_or("".to_string(), |m| m.to_string()),
-    );
-    let guest_error_buffer: Vec<u8> = (&guest_error)
-        .try_into()
-        .expect("Invalid guest_error_buffer, could not be converted to a Vec<u8>");
-
-    push_shared_output_data(guest_error_buffer)
-        .expect("Unable to push guest error to shared output data");
-}
-
-pub(crate) fn set_error(error_code: ErrorCode, message: &str) {
-    write_error(error_code, Some(message));
-}
-
+// TODO(danbugs): move this to `hyperlight_guest_bin`
 /// Exposes a C API to allow the guest to set an error
 ///
 /// # Safety
@@ -48,14 +30,17 @@ pub(crate) fn set_error(error_code: ErrorCode, message: &str) {
 #[unsafe(no_mangle)]
 #[allow(non_camel_case_types)]
 pub unsafe extern "C" fn setError(code: u64, message: *const c_char) {
+    let handle = unsafe { GUEST_HANDLE };
+
     let error_code = ErrorCode::from(code);
     match message.is_null() {
-        true => write_error(error_code, None),
+        true => handle.write_error(error_code, None),
         false => {
             let message = unsafe { CStr::from_ptr(message).to_str().ok() }
                 .expect("Invalid error message, could not be converted to a string");
-            write_error(error_code, Some(message));
+            handle.write_error(error_code, Some(message));
         }
     }
+
     halt();
 }
