+++ src/hyperlight_guest_bin/src/guest_function/call.rs
@@ -20,13 +20,10 @@ use alloc::vec::Vec;
 use hyperlight_common::flatbuffer_wrappers::function_call::{FunctionCall, FunctionCallType};
 use hyperlight_common::flatbuffer_wrappers::function_types::ParameterType;
 use hyperlight_common::flatbuffer_wrappers::guest_error::ErrorCode;
+use hyperlight_guest::error::{HyperlightGuestError, Result};
+use hyperlight_guest::exit::halt;
 
-use crate::entrypoint::halt;
-use crate::error::{HyperlightGuestError, Result};
-use crate::guest_error::set_error;
-use crate::shared_input_data::try_pop_shared_input_data_into;
-use crate::shared_output_data::push_shared_output_data;
-use crate::REGISTERED_GUEST_FUNCTIONS;
+use crate::{GUEST_HANDLE, REGISTERED_GUEST_FUNCTIONS};
 
 type GuestFunc = fn(&FunctionCall) -> Result<Vec<u8>>;
 
@@ -83,17 +80,20 @@ pub(crate) fn call_guest_function(function_call: FunctionCall) -> Result<Vec<u8>
 #[unsafe(no_mangle)]
 #[inline(never)]
 fn internal_dispatch_function() -> Result<()> {
+    let handle = unsafe { GUEST_HANDLE };
+
     #[cfg(debug_assertions)]
     log::trace!("internal_dispatch_function");
 
-    let function_call = try_pop_shared_input_data_into::<FunctionCall>()
+    let function_call = handle
+        .try_pop_shared_input_data_into::<FunctionCall>()
         .expect("Function call deserialization failed");
 
     let result_vec = call_guest_function(function_call).inspect_err(|e| {
-        set_error(e.kind.clone(), e.message.as_str());
+        handle.write_error(e.kind.clone(), Some(e.message.as_str()));
     })?;
 
-    push_shared_output_data(result_vec)
+    handle.push_shared_output_data(result_vec)
 }
 
 // This is implemented as a separate function to make sure that epilogue in the internal_dispatch_function is called before the halt()
